\name{ex_alignment1}
\alias{ex_alignment1}
\title{example alignment #1}
\description{This function will generate an example alignment to test some function within this package}
\usage{data(ex_alignment1)}
\details{
Because fasta file examples are not automatically loaded into R environment, 'ex_alignment1' function generates a fasta file (named Example_Spatial.plot_Alignment) that is stored as a 'DNAbin' object named alignExample
}

\author{A. J. Muñoz-Pajares}

\seealso{
\code{\link{alignExample}}, \code{\link{Example_Spatial.plot_Alignment}}
}
\examples{
# data(ex_alignment1) # this will read a fasta file with the name 'alignExample'
# alignExample
}
