\name{ex_BLAST}
\alias{ex_BLAST}
\title{example BLAST output}
\description{BLAST output}
\usage{data(ex_BLAST)}
\value{a four columns data.frame containing the qseqid (as character, not as factor!), evalue, percentage of identity and query coverage, respectively.}

\author{A. J. Muñoz-Pajares}

\examples{
# data(ex_BLAST)
# ex_BLAST
}
