\name{ex_Coords}
\alias{ex_Coords}
\title{example coordinates}
\description{Geographic coordinates of populatons to test some function within this package}
\usage{data(ex_Coords)}
\value{a three columns matrix containing the population name, longitude and latitude, respectively.}

\author{A. J. Muñoz-Pajares}

\examples{
# data(ex_Coords)
# ex_Coords
}
