\name{Example_Spatial.plot_Alignment}
\alias{Example_Spatial.plot_Alignment}
\title{example alignment #1 (fasta format)}
\description{fasta sequences alignment to test some function within this package}
\usage{data(ex_alignment1)}
\details{
Because fasta file examples are not automatically loaded into R environment, 'ex_alignment1' function generates a fasta file (named Example_Spatial.plot_Alignment) that is stored as a 'DNAbin' object named alignExample
}

\author{A. J. Muñoz-Pajares}

\seealso{
\code{\link{ex_alignment1}}, \code{\link{alignExample}}
}
\examples{
# data(ex_alignment1) # this will read a fasta file with the name 'alignExample'
# alignExample
}
