\name{genbank.sp.names}
\alias{genbank.sp.names}
\title{Species names from genbank accessions
}
\description{
Given a set of sequences in fasta format, this function extract the species names of every accession.
}
\usage{
genbank.sp.names(sequences)
}
\arguments{
  \item{sequences}{
a DNAbin object containing all the information of genbank accesions as sequence names.
}

}
\value{
a DNAbin object with genus and species as sequence names.
}

\author{
A.J. Muñoz-Pajares
}


