\name{sidier-package}
\alias{sidier-package}
\alias{sidier}
\docType{package}
\title{SIDIER: Substitution and Indel Distances to Infer Evolutionary Relationships
}
\description{sidier is a library and R package for evolutionary reconstruction based on substitutions and insertion-deletion (indels) analyses in a distance-based framework.
}
\details{
\tabular{ll}{
Package: \tab sidier\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-02-16\cr
License: \tab GPL-2\cr
}
}
\author{A. Jesus Munoz Pajares\\
Maintainer: A. Jesus Munoz Pajares <ajesusmp@ugr.es>}
\references{
Muller K. (2006). Incorporating information from length-mutational events into phylogenetic analysis. Molecular Phylogenetics and Evolution, 38, 667-676.

Munoz-Pajares, AJ. SIDIER: Substitution and Indel Distances to Infer Evolutionary Relationships.

Munoz-Pajares, A.J., Abdelaziz, M., Gomez, J.M., Perfectti, F. Combining indels and substitutions information for the reconstruction of evolutionary haplotype relationships.

Munoz-Pajares, A.J., Abdelaziz, M., Herrador, M.B., Gomez, J.M., Perfectti, F. Phylogeography and colonization pathways of the Erysimum nevadense species complex based on a plastidial indel-rich region distance analysis.

Rozenfeld AF, Arnaud-Haond S, Hernandez-Garcia E, Eguiluz VM, Serrao EA, Duarte CM. (2008). Network analysis identifies weak and strong links in a metapopulation system. Proceedings of the National Academy of Sciences,105, 18824-18829.

}
