\name{sidier-package}
\alias{sidier-package}
\alias{sidier}
\docType{package}
\title{SIDIER: Substitution and Indel Distances to Infer Evolutionary Relationships}
\description{sidier is a library and R package for evolutionary reconstruction based on substitutions and insertion-deletion (indels) analyses in a distance-based framework.}
\details{
\tabular{ll}{
Package: \tab sidier\cr
Type: \tab Package\cr
Version: \tab 3.0.1\cr
Date: \tab 2015-02-27\cr
License: \tab GPL-2\cr
}

This package estimates evolutionary distances based on indels that may be combined with the information provided by substitutions and depicted as networks.
}
\author{
A. Jesús Muñoz Pajares

Maintainer:  A. J. Muñoz-Pajares <ajesusmp@ugr.es>
}
\references{
Muñoz-Pajares, A. J. (2013). SIDIER: substitution and indel distances to infer evolutionary relationships. Methods in Ecology and Evolution 4, 1195-1200. doi: 10.1111/2041-210X.12118
}

