\name{sidier-package}
\alias{sidier-package}
\alias{sidier}
\docType{package}
\title{SIDIER: Substitution and Indel Distances to Infer Evolutionary Relationships
}
\description{sidier is a library and R package for evolutionary reconstruction based on substitutions and insertion-deletion (indels) analyses in a distance-based framework.
}
%\details{
%\tabular{ll}{
%Package: \tab sidier\cr
%Type: \tab Package\cr
%Version: \tab 2.2\cr
%Date: \tab 2013-06-30\cr
%License: \tab GPL-2\cr
%Depends: \tab R (>= 2.10.1) ape, network, igraph\cr
%Imports: \tab ape, network, igraph\cr
%}
%~~ An overview of how to use the package, including the most important functions ~~
%}
\author{A. Jesus Munoz Pajares}%\cr
%\cr
%Maintainer: A. Jesus Munoz Pajares <ajesusmp@ugr.es>}
\references{

Munoz-Pajares, AJ. (Submitted) SIDIER: Substitution and Indel Distances to Infer Evolutionary Relationships.

}
